Move transfer parsing function out of module
